refactor(forms): clean up interest/app processors
Refactor the interest and application form processors, and make other
minor improvements.

* Add missing "defer tx.Rollback()" before activist insert queries
* Remove unnecessary log lines in happy paths
* Split large functions; return errors instead of logging directly
* Split interest/application logic into separate files
* Move SQL queries into related files
